requirements.txt
openai
faiss-cpu
python-docx
numpy
scipy
python-dotenv
dspy-ai